A2A client refactor + Makefile  (#185)
* Move save function to io_utils

* Break down test_client into smaller functions

* Rename test_client -> agent_generation

* Introduce Makefile

* Change description

* Address review comments: Makefile and readme

* Update error condition of stop target in makefile

* Rename to agent_generator
