+++ src/agent_factory/utils/__init__.py
@@ -1,9 +1,13 @@
 from .artifact_validation import clean_python_code_with_autoflake
-from .io_utils import setup_output_directory
-from .logging import logger
+from .client_utils import create_a2a_http_client, create_message_request, get_a2a_agent_card, process_a2a_agent_response
+from .io_utils import save_agent_outputs, setup_output_directory
 
 __all__ = [
     "clean_python_code_with_autoflake",
     "setup_output_directory",
-    "logger",
+    "save_agent_outputs",
+    "create_a2a_http_client",
+    "get_a2a_agent_card",
+    "create_message_request",
+    "process_a2a_agent_response",
 ]
